requirements.txt
apscheduler==3.10.4
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
pyyaml==6.0.1
pytz==2024.1
anthropic==0.28.0
python-dotenv==1.0.1
